# Use SQLITE transactions when inserting PCAP data in `recap`.