--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -59,6 +59,6 @@ set_target_properties(edgesec PROPERTIES INTERPROCEDURAL_OPTIMIZATION TRUE)
 
 if (USE_CAPTURE_SERVICE)
   add_executable(recap recap.c)
-  target_link_libraries(recap PRIVATE protobuf_middleware packet_queue packet_decoder sqlite_header os log SQLite::SQLite3)
+  target_link_libraries(recap PRIVATE protobuf_middleware packet_queue packet_decoder sqlite_header os log SQLite::SQLite3 sqliteu)
   target_include_directories(recap PRIVATE ${PROJECT_BINARY_DIR})
 endif()