Modifies the recap executable to pipe all of the PCAP data into the sqlite database using a single SQLITE transaction.

This has two benefits:

1. On an error, the entire transaction is rolled-back, so we won't have a half imported recap file.
2. Running recap is much, much, much, much faster.
   **A quick local test went from 12.0KiB/s to 40.4 MiB/s, which is a ~3300x increase in speed!**

### Benchmarks

When using the `cmake --preset linux` setting:

#### Before (commit 12a7aa444da23f8225288aafe67dda0696f5d741)

After running for 5 minutes, it has only imported 2 MiB of the PCAP file.

```console
alois@laptop:~/Documents/edgesec ((12a7aa44...))$ timeout 5m bash -c 'pv ~/Downloads/2021_11_24_Active.pcap | ./build/linux/src/recap -f /tmp/recap-pcap-benchmark.sqlite'
Opened db at /tmp/recap-pcap-benchmark.sqlite
1.98MiB 0:04:59 [12.0KiB/s] [>                              ]  0% ETA 9:19:41:17
```

#### After (commit ce3c5f5b93f27ac864997f8f8cdc243e0caa23c8)

After running for 2:18 minutes, it had imported 5.48 GiB of the PCAP file (i.e. all of it).

Interestingly, it looks like it's now bottlenecked by CPU speed.

```console
alois@laptop:~/Documents/edgesec (optimize-recap)$ timeout 5m bash -c 'pv ~/Downloads/2021_11_24_Active.pcap | ./build/linux/src/recap -f /tmp/recap-pcap-benchmark.sqlite'
Opened db at /tmp/recap-pcap-benchmark.sqlite
5.48GiB 0:02:18 [40.4MiB/s] [================================>] 100%            
Processed pcap size = 5883456163 bytes
Processed packets = 18555332
```

I also tried compiling with `-DCMAKE_BUILD_TYPE=Release` (e.g. build with optimizations) and it actually seemed to slow things down to 2:40 minutes, but maybe that's just because my laptop has a terrible quality heatsink and it was fully saturated :laughing:

### Notes

I've also done a bit of minor refactoring while I was in the function, so please review this PR commit by commit!

It's up to you if this closes https://github.com/nqminds/edgesec/pull/347.
The optimizations in that PR should be cumulative with this PR, but it is at the cost of adding additional danger.