@aloisklink 2022-11-21
> Very good. There's one option when immediate transaction are not desirable, i.e., the case of continuous data ingestion from the libpcap stream. Currently I'm working on a branch that adds libpcap to recap. So ideally we need to add a flag "-u" (or similar) that allows user to use the commit functionality and this functionality is diasble when a continuous stream is used.

Added in https://github.com/nqminds/edgesec/pull/348/commits/dad2b981e7b287e6d94fccbb75f9430b89d0e599!

This is what the new `recap -h` looks like:

```console
alois@laptop:~/Documents/edgesec (optimize-recap)$ ./build/linux/src/recap -h
recap app version 0.1.0-alpha.0
Usage:
	recap [-p filename] [-f filename] [-i interface] [-k] [-d] [-h] [-v]
		 [-t {SINGLE_TRANSACTION,DISABLED}]
Run capture on an input pcap file and output to a capture db or pipe.

Options:
	-p filename	 Path to the pcap file name
	-f filename	 Path to the capture db or pipe
	-i interface	 Interface name to save to db
	-t TRANS_TYPE	 Controls how SQLITE transactions are used.
	             	 Allowed values:
	             	   SINGLE_TRANSACTION: Use a single SQLITE transaction.
	             	   	 Provides maximum insertion performance.
	             	   DISABLED: Don't use SQLITE transactions (very slow).
	             	   (default) Pick automatically based on input.
	-k		 Pipe to file
	-d		 Verbosity level (use multiple -dd... to increase)
	-h		 Show help
	-v		 Show app version

Copyright NQMCyber Ltd
```

Basically, you can pass `-t SINGLE_TRANSACTION` to add everything in a super fast single transaction, `-t DISABLED` to be 3000x slower, or leave it blank, and `recap` will automatically choose an option for you (e.g. can be disabled for continuous streams).

I've also added some basic tests in https://github.com/nqminds/edgesec/commit/c4a245d1d90995c5495ba869ec70b7e1d5a93417 for the following cases in recap:

  - `recap -v`
  - `recap -f tmpdb -p pcap`
  - `recap -f tmpdb -p pcap -t SINGLE_TRANSACTION`
    (insert pcap into sqlite db in a single transaction)
  - `recap -f tmpdb -p pcap -t DISABLED`
    (insert pcap into sqlite db without transactions) (**warning, this takes like 5 seconds to complete, so it turns the original 2.5 second tests to 7.5 seconds**)
  - `recap -f tmpdb -p pcap -t invalid-arg-here`
    (checks for specific error message)
  - cleanup tmpdb when all tests are done